requirements.txt
pyside
pyinstaller
